Stop using peribolos for managing org membership (#226)
* Remove github action to apply org membership

We are no longer going to use Peribolos to manage permissions in the ODH
org. Removing the action for applying membership so that changes don't
get applied.

* Stop using peribolos for managing org membership

We've agreed at an ODH leadership level that we no longer want to use
Peribolos for managing membership in the ODH org. This change removes
the peribolos-specific contents of this repository and updates the
README accordingly.